Parameters used in multiple functions

Reused parameter definitions are gathered here for easier usage.

Arguments

id

The namespace id for this module.

combinedData

data.frame A combined data set with the required metrics.

dfSite

data.frame Optional site-level metadata.

domain

The type of data and subsequent metrics you would like to see.

input

Shiny inputs

lAssessment

list List of KRI assessments from params within KRIReport.Rmd.

metric

The reactive value provided by the metric input from server

output

Shiny outputs

session

Shiny session

site

The reactive value provided by the site input from server

snapshot

A list object containing the data for a snapshot. This object funnels into virtually every function of the application.

subset_col

The id column that will be selected from the domain.

subset_val

The specific value that the subset_col will be filtered by.